requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
crispy-bootstrap4
Django
django-crispy-forms
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
idna
inflection
itypes
Jinja2
MarkupSafe
packaging
pillow
PyJWT
pytz
PyYAML
requests
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
whitenoise